backend/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
asgiref==3.11.1
attrs==25.4.0
backoff==2.2.1
bcrypt==4.1.3
black==26.1.0
boto3==1.35.89
botocore==1.35.99
build==1.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.3.1
cryptography==46.0.5
distro==1.9.0
dnspython==2.8.0
durationpy==0.10
ecdsa==0.19.1
email-validator==2.3.0
emergentintegrations==0.1.0
et_xmlfile==2.0.0
fastapi==0.110.1
fastuuid==0.14.0
filelock==3.24.3
flake8==7.3.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.64.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.1
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.3.1
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
isort==8.0.0
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
jq==1.11.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
librt==0.8.1
litellm==1.80.0
lxml==5.4.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.2.0
motor==3.3.1
mpmath==1.3.0
multidict==6.7.1
mypy==1.19.1
mypy_extensions==1.1.0
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.24.2
openai==1.99.9
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
overrides==7.7.0
packaging==26.0
pandas==2.3.3
passlib==1.7.4
pathspec==1.0.4
pillow==10.4.0
platformdirs==4.9.2
pluggy==1.6.0
posthog==7.9.4
prometheus_client==0.24.1
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
pulsar-client==3.10.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.11.0
pymongo==4.5.0
PyMuPDF==1.27.1
pyotp==2.9.0
pyparsing==3.3.2
PyPDF2==3.0.1
PyPika==0.51.1
pypng==0.20220715.0
pyproject_hooks==1.2.0
pytest==9.0.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
pytokens==0.4.1
pytz==2026.1.post1
PyYAML==6.0.3
qrcode==7.4.2
redis==5.0.1
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
requests-oauthlib==2.0.0
resend==2.23.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.10.4
s5cmd==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
stripe==14.4.1
sympy==1.14.0
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.25.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
